requirements.txt
fastapi 
uvicorn 
sqlalchemy 
asyncpg 
alembic
pydantic-settings 
pydantic[email]
python-jose[cryptography] 
passlib[bcrypt]
redis 
celery 
slowapi